backend/requirements.txt
langchain_openai
langgraph
fastapi
uvicorn
langchain
langchain-openai
pydantic
typing-extensions
requests
langchain-google-genai
google-generativeai
dotenv